+++ CHANGELOG.md
@@ -98,6 +98,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - xDnsServerSetting
   - Changed to use `Get-DnsServerSetting` and `Set-DnsServerSetting`
     ([issue #185](https://github.com/dsccommunity/xDnsServer/issues/185)).
+    - Property `DisableAutoReverseZone` was renamed from `DisableAutoReverseZones`.
 
 ### Fixed